[refactor] attendancecheck 패키지 이름 표준 패턴으로 변경 (#80)
