--- src/main/java/konkuk/thip/attendanceCheck/adapter/in/web/AttendanceCheckCommandController.java
@@ -1,10 +1,0 @@
-package konkuk.thip.attendanceCheck.adapter.in.web;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class AttendanceCheckCommandController {
-
-}
