--- src/main/java/konkuk/thip/attendanceCheck/adapter/in/web/request/UserSignupRequest.java
@@ -1,7 +1,0 @@
-package konkuk.thip.attendanceCheck.adapter.in.web.request;
-
-import lombok.Getter;
-
-@Getter
-public class UserSignupRequest {
-}
