--- src/main/java/konkuk/thip/attendanceCheck/adapter/in/web/response/DummyResponse.java
@@ -1,7 +1,0 @@
-package konkuk.thip.attendanceCheck.adapter.in.web.response;
-
-import lombok.Getter;
-
-@Getter
-public class DummyResponse {
-}
