--- src/main/java/konkuk/thip/attendanceCheck/application/port/in/DummyUseCase.java
@@ -1,5 +1,0 @@
-package konkuk.thip.attendanceCheck.application.port.in;
-
-public interface DummyUseCase {
-
-}
