--- src/main/java/konkuk/thip/attendanceCheck/application/port/in/dto/DummyCommand.java
@@ -1,10 +1,0 @@
-package konkuk.thip.attendanceCheck.application.port.in.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class DummyCommand {
-
-}
