--- src/main/java/konkuk/thip/attendanceCheck/application/port/in/dto/DummyQuery.java
@@ -1,9 +1,0 @@
-package konkuk.thip.attendanceCheck.application.port.in.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class DummyQuery {
-}
