--- src/main/java/konkuk/thip/attendanceCheck/application/port/in/dto/DummyResult.java
@@ -1,10 +1,0 @@
-package konkuk.thip.attendanceCheck.application.port.in.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class DummyResult {
-
-}
