--- src/main/java/konkuk/thip/attendanceCheck/application/service/AttendanceCheckService.java
@@ -1,11 +1,0 @@
-package konkuk.thip.attendanceCheck.application.service;
-
-import konkuk.thip.attendanceCheck.application.port.in.DummyUseCase;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AttendanceCheckService implements DummyUseCase {
-
-}
