--- src/main/java/konkuk/thip/attendanceCheck/domain/AttendanceCheck.java
@@ -1,18 +1,0 @@
-package konkuk.thip.attendanceCheck.domain;
-
-import konkuk.thip.common.entity.BaseDomainEntity;
-import lombok.Getter;
-import lombok.experimental.SuperBuilder;
-
-@Getter
-@SuperBuilder
-public class AttendanceCheck extends BaseDomainEntity {
-
-    private Long id;
-
-    private String todayComment;
-
-    private Long roomId;
-
-    private Long creatorId;
-}
