+++ src/main/java/konkuk/thip/attendancecheck/adapter/in/web/AttendanceCheckQuerydController.java
@@ -1,0 +1,10 @@
+package konkuk.thip.attendancecheck.adapter.in.web;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class AttendanceCheckQuerydController {
+
+}
