+++ src/main/java/konkuk/thip/attendancecheck/adapter/in/web/request/UserSignupRequest.java
@@ -1,0 +1,7 @@
+package konkuk.thip.attendancecheck.adapter.in.web.request;
+
+import lombok.Getter;
+
+@Getter
+public class UserSignupRequest {
+}
