+++ src/main/java/konkuk/thip/attendancecheck/adapter/in/web/response/DummyResponse.java
@@ -1,0 +1,7 @@
+package konkuk.thip.attendancecheck.adapter.in.web.response;
+
+import lombok.Getter;
+
+@Getter
+public class DummyResponse {
+}
