+++ src/main/java/konkuk/thip/attendancecheck/application/port/in/DummyUseCase.java
@@ -1,0 +1,5 @@
+package konkuk.thip.attendancecheck.application.port.in;
+
+public interface DummyUseCase {
+
+}
