+++ src/main/java/konkuk/thip/attendancecheck/application/port/in/dto/DummyCommand.java
@@ -1,0 +1,10 @@
+package konkuk.thip.attendancecheck.application.port.in.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class DummyCommand {
+
+}
