+++ src/main/java/konkuk/thip/attendancecheck/application/port/in/dto/DummyQuery.java
@@ -1,0 +1,9 @@
+package konkuk.thip.attendancecheck.application.port.in.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class DummyQuery {
+}
