+++ src/main/java/konkuk/thip/attendancecheck/application/service/AttendanceCheckService.java
@@ -1,0 +1,11 @@
+package konkuk.thip.attendancecheck.application.service;
+
+import konkuk.thip.attendancecheck.application.port.in.DummyUseCase;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AttendanceCheckService implements DummyUseCase {
+
+}
